optional SSL for usyncd
usyncd now only enables a SSL context if found in the settings, in more
realistic scenarios it should run behind a proxy.

Moved client.py uri variable up to make it more obvious a user defined
value. This can currently be tested with either a local instance or the
usync.org example.

Signed-off-by: Paul Spooren <mail@aparcar.org>